+++ packages/outbox/src/test/utils/helpers/shared/logger-mock.helper.ts
@@ -1,9 +1,8 @@
-import type { jest } from '@jest/globals';
 import type { Logger } from '@volontariapp/logger';
-import { createMock } from '@volontariapp/database/testing';
+import { createMockLogger } from '@volontariapp/testing';
 
 export type LoggerMock = jest.Mocked<Logger>;
 
 export const makeLoggerMock = (): LoggerMock => {
-  return createMock<Logger>();
+  return createMockLogger<Logger>();
 };