+++ packages/domain-social/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.3.13
+
+### Patch Changes
+
+- testing lib added
+
+- Updated dependencies []:
+  - @volontariapp/logger@0.2.4
+  - @volontariapp/bridge-nest@0.3.13
+  - @volontariapp/contracts@4.0.10
+  - @volontariapp/errors@0.5.1
+  - @volontariapp/errors-nest@0.11.1
+
 ## 0.3.12
 
 ### Patch Changes