+++ packages/errors/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/logger@0.2.4
+
 ## 0.5.0
 
 ### Minor Changes